+++ src/app/components/sidebar/sidebar.component.ts
@@ -308,6 +308,7 @@ export class SidebarComponent implements OnInit, AfterViewInit {
   PERMISSION_TO_VIEW_KB: boolean;
   PERMISSION_TO_VIEW_ANALYTICS: boolean;
   PERMISSION_TO_VIEW_ACTVITIES: boolean;
+  PERMISSION_TO_VIEW_DATATABLES: boolean;
   PERMISSION_TO_VIEW_WA_BRODCAST: boolean;
   PERMISSION_TO_VIEW_SETTING: boolean;
   isTiledeskDomain = false;
@@ -331,6 +332,7 @@ export class SidebarComponent implements OnInit, AfterViewInit {
   PERMISSION_TO_VIEW_PROJECT_SETTING_NOTIFICATION: boolean;
   PERMISSION_TO_VIEW_PROJECT_SETTING_SECURITY: boolean;
   PERMISSION_TO_VIEW_PROJECT_SETTING_BANNED: boolean;
+  PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION: boolean;
   PERMISSION_TO_VIEW_PROJECT_SETTING_ADVANCED: boolean;
   PERMISSION_TO_VIEW_HOME: boolean;
   
@@ -580,6 +582,25 @@ export class SidebarComponent implements OnInit, AfterViewInit {
           // Custom roles: permission depends on matchedPermissions
           this.PERMISSION_TO_VIEW_ACTVITIES = status.matchedPermissions.includes(PERMISSIONS.ACTIVITIES_READ);
           this.logger.log('[SIDEBAR] - Custom role (3) role', status.role, 'PERMISSION_TO_VIEW_ACTVITIES:', this.PERMISSION_TO_VIEW_ACTVITIES);
+        }
+
+        // -----------------------------
+        // PERMISSION_TO_VIEW_DATATABLES
+        // -----------------------------
+         if (status.role === 'owner' || status.role === 'admin') {
+          // Owner and admin always has permission
+          this.PERMISSION_TO_VIEW_DATATABLES = true;
+          this.logger.log('[SIDEBAR] - Project user is owner or admin (1)', 'PERMISSION_TO_VIEW_DATATABLES:', this.PERMISSION_TO_VIEW_DATATABLES);
+
+        } else if (status.role === 'agent') {
+          // Agent never have permission
+          this.PERMISSION_TO_VIEW_DATATABLES = false;
+          this.logger.log('[SIDEBAR] - Project user agent (2)', 'PERMISSION_TO_VIEW_DATATABLES:', this.PERMISSION_TO_VIEW_DATATABLES);
+
+        } else {
+          // Custom roles: permission depends on matchedPermissions
+          this.PERMISSION_TO_VIEW_DATATABLES = status.matchedPermissions.includes(PERMISSIONS.DATATABLES_READ);
+          this.logger.log('[SIDEBAR] - Custom role (3) role', status.role, 'PERMISSION_TO_VIEW_DATATABLES:', this.PERMISSION_TO_VIEW_DATATABLES);
         }
 
         // -------------------------------
@@ -852,6 +873,7 @@ export class SidebarComponent implements OnInit, AfterViewInit {
                   PERMISSIONS.PROJECTSETTINGS_NOTIFICATION_READ,
                   PERMISSIONS.PROJECTSETTINGS_SECURITY_READ,
                   PERMISSIONS.PROJECTSETTINGS_BANNED_READ,
+                  PERMISSIONS.PROJECTSETTINGS_RETENTION_READ,
                   PERMISSIONS.PROJECTSETTINGS_ADVANCED_READ
               ];
               
@@ -975,6 +997,25 @@ export class SidebarComponent implements OnInit, AfterViewInit {
             // Custom roles: permission depends on matchedPermissions
             this.PERMISSION_TO_VIEW_PROJECT_SETTING_BANNED = status.matchedPermissions.includes(PERMISSIONS.PROJECTSETTINGS_BANNED_READ);
             this.logger.log('[SIDEBAR] - Custom role (3) role', status.role, 'PERMISSION_TO_VIEW_PROJECT_SETTING_BANNED:', this.PERMISSION_TO_VIEW_PROJECT_SETTING_BANNED);
+          }
+
+          // --------------------------------------------------
+          // PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION
+          // -------------------------------------------------
+          if (status.role === 'owner' || status.role === 'admin') {
+            // Owner and admin always has permission
+            this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION = true;
+            this.logger.log('[SIDEBAR] - Project user is owner or admin (1)', 'PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION:', this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION);
+
+          } else if (status.role === 'agent') {
+            // Agent never have permission
+            this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION = false;
+            this.logger.log('[SIDEBAR] - Project user agent (2)', 'PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION:', this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION);
+
+          } else {
+            // Custom roles: permission depends on matchedPermissions
+            this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION = status.matchedPermissions.includes(PERMISSIONS.PROJECTSETTINGS_RETENTION_READ);
+            this.logger.log('[SIDEBAR] - Custom role (3) role', status.role, 'PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION:', this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION);
           }
 
           // --------------------------------------------------
@@ -2955,6 +2996,10 @@ export class SidebarComponent implements OnInit, AfterViewInit {
 
       this.goToProjectSettings_BannedVisitors()
 
+    } else if (this.PERMISSION_TO_VIEW_PROJECT_SETTING_RETENTION) {
+
+      this.goToProjectSettings_Retention()
+
     } else if (this.PERMISSION_TO_VIEW_PROJECT_SETTING_ADVANCED)  {
         
       this.goToProjectSettings_Advanced()
@@ -2988,6 +3033,10 @@ export class SidebarComponent implements OnInit, AfterViewInit {
 
   goToProjectSettings_BannedVisitors() {
     this.router.navigate(['project/' + this.project._id + '/project-settings/banned'])
+  }
+
+  goToProjectSettings_Retention() {
+    this.router.navigate(['project/' + this.project._id + '/project-settings/retention'])
   }
 
   goToProjectSettings_Advanced() {