+++ src/app/integrations/list/hubspot-integration/hubspot-integration.component.ts
@@ -14,10 +14,9 @@ export class HubspotIntegrationComponent implements OnInit {
   @Output() onUpdateIntegration = new EventEmitter;
   @Output() onDeleteIntegration = new EventEmitter;
 
-  keyVisibile: boolean = false;
   isVerified: boolean;
   translateparams: any;
-  isMasked: boolean = true; // State for masking
+  apiKeyCanSave = false;
   
   constructor(
     private integrationService: IntegrationService,
@@ -31,17 +30,6 @@ export class HubspotIntegrationComponent implements OnInit {
     // if (this.integration.value.apikey) {
     //   this.checkKey();
     // }
-  }
-
-  showHideKey() {
-    this.logger.log("showHideKey called");
-    let input = <HTMLInputElement>document.getElementById('api-key-input');
-    if (this.keyVisibile === false) {
-      input.type = 'text';
-    } else {
-      input.type = 'password';
-    }
-    this.keyVisibile = !this.keyVisibile;
   }
 
   saveIntegration() {
@@ -118,43 +106,4 @@ export class HubspotIntegrationComponent implements OnInit {
       apikey: null
     }
   }
-  // ---------------------------------------------------
-  // Mask Api key without use input of password type
-  // ---------------------------------------------------
-  handleInput(event: Event): void {
-    const inputElement = event.target as HTMLInputElement;
-    const displayedValue = inputElement.value;
-
-    // Update realValue based on input length and masking state
-    if (this.isMasked && this.integration.value.apikey) {
-      // Add only new characters to realValue
-      const newChar = displayedValue.slice(this.integration.value.apikey.length);
-      this.integration.value.apikey += newChar;
-    } else {
-      // Directly update realValue when unmasked
-      this.integration.value.apikey = displayedValue;
-    }
-
-    // Always set the displayed value to match the current state
-    inputElement.value = this.getDisplayValue();
-  }
-
-  handleBackspace(): void {
-    this.integration.value.apikey = this.integration.value.apikey.slice(0, -1);
-  }
-
-  toggleMask(inputElement: HTMLInputElement): void {
-    this.isMasked = !this.isMasked;
-
-    // Update the displayed value immediately when toggling the mask
-    inputElement.value = this.getDisplayValue();
-  }
-
-  getDisplayValue(): string {
-    if (!this.integration.value.apikey) {
-      return ''; // Return an empty string if realValue is null, undefined, or empty
-    }
-    return this.isMasked ? '●'.repeat(this.integration.value.apikey.length) : this.integration.value.apikey;
-  }
-
 }